+++ Editor/Sources/Recorder.cs
@@ -23,7 +23,10 @@ namespace UnityEditor.Recorder
     {
         static int sm_CaptureFrameRateCount;
         bool m_ModifiedCaptureFR;
-        bool m_asyncShaderCompileSetting;
+
+        private static bool s_asyncShaderCompileSetting;
+        private static bool s_asyncShaderCompileAlreadyRestored = false; // have we already restored the value of the setting?
+        private static bool s_asyncShaderCompileAlreadyDisabled = false; // have we already disabled the setting?
 
         /// <summary>
         /// Indicates the number of frames of the current recording session.
@@ -168,10 +171,7 @@ namespace UnityEditor.Recorder
             if (RecorderOptions.VerboseMode)
                 ConsoleLogMessage($"Starting to record", LogType.Log);
 
-            // Save the async compile shader setting to restore it at the end of recording
-            m_asyncShaderCompileSetting = EditorSettings.asyncShaderCompilation;
-            // Disable async compile shader setting when recording
-            EditorSettings.asyncShaderCompilation = false;
+            DisableAsyncShaderCompil();
 
             return Recording = true;
         }
@@ -208,9 +208,7 @@ namespace UnityEditor.Recorder
             if (RecorderOptions.VerboseMode)
                 ConsoleLogMessage($"Recording stopped, total frame count: {RecordedFramesCount}", LogType.Log);
 
-            // Restore the asyncShaderCompilation setting
-            EditorSettings.asyncShaderCompilation = m_asyncShaderCompileSetting;
-
+            RestoreAsynchronousShaderCompilation();
             ++settings.Take;
         }
 
@@ -322,6 +320,34 @@ namespace UnityEditor.Recorder
                 default:
                     throw new ArgumentOutOfRangeException("stage", stage, null);
             }
+        }
+
+        /// <summary>
+        /// Disable asynchronous shader compilation and save the status.
+        /// </summary>
+        private static void DisableAsyncShaderCompil()
+        {
+            if (s_asyncShaderCompileAlreadyDisabled)
+                return;
+
+            // Save the async compile shader setting to restore it at the end of recording
+            s_asyncShaderCompileSetting = EditorSettings.asyncShaderCompilation;
+            // Disable async compile shader setting when recording
+            EditorSettings.asyncShaderCompilation = false;
+            s_asyncShaderCompileAlreadyRestored = false;
+            s_asyncShaderCompileAlreadyDisabled = true;
+        }
+
+        /// <summary>
+        /// If we have not already restored the setting of asynchronous shader compilation, restore it.
+        /// </summary>
+        private static void RestoreAsynchronousShaderCompilation()
+        {
+            if (s_asyncShaderCompileAlreadyRestored || !s_asyncShaderCompileAlreadyDisabled)
+                return;
+            EditorSettings.asyncShaderCompilation = s_asyncShaderCompileSetting;
+            s_asyncShaderCompileAlreadyRestored = true;
+            s_asyncShaderCompileAlreadyDisabled = false;
         }
     }
 }