+++ Editor/Sources/RecorderWindow.cs
@@ -60,6 +60,12 @@ namespace UnityEditor.Recorder
             }
         }
 
+        [MenuItem(MenuRoot + "Quick Recording _F10", true, MenuRootIndex + 1)]
+        static bool QuickRecordingValidate()
+        {
+            return !EditorUtility.scriptCompilationFailed;
+        }
+
         static void Progress()
         {
             if (s_NeedToResetView)
@@ -510,6 +516,11 @@ namespace UnityEditor.Recorder
         /// </summary>
         public void StartRecording()
         {
+            if (EditorUtility.scriptCompilationFailed)
+            {
+                return;
+            }
+
             if (EditorApplication.isPlaying)
             {
                 // Already in play mode, so start recording now
@@ -858,7 +869,7 @@ namespace UnityEditor.Recorder
                 }
                 else
                 {
-                    SetRecordButtonsEnabled(true);
+                    SetRecordButtonsEnabled(!EditorUtility.scriptCompilationFailed);
                 }
             }
             else