+++ Editor/Timeline/_WaitForEndOfFrameComponent.cs
@@ -3,6 +3,10 @@ using UnityEngine;
 
 namespace UnityEditor.Recorder.Timeline
 {
+    // Hack to make new a new MonoBehaviour discovery feature in the engine.
+    // This is because there is a mismatched between the MB name and the file name, all inside an Editor assembly (double nono).
+    class _WaitForEndOfFrameComponent{}
+
     [ExecuteInEditMode]
     class WaitForEndOfFrameComponent : _FrameRequestComponent
     {