+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [3.0.3] - 2021-11-08
+### Added
+### Changed
+### Fixed
+- Prevent the user from starting a recording session from the Recorder window if the Unity Editor has compiler errors.
+- Make sure to always restore the asynchronous shader compilation setting's original value when the recording session ends.
+
 ## [3.0.2] - 2021-09-09
 ### Added
 ### Changed